requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
black==22.3.0
    # via -r requirements.in
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
decorator==5.1.1
    # via ipython
django==4.0.5
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-filter
    #   djangorestframework
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
executing==0.8.3
    # via stack-data
idna==3.3
    # via requests
ipython==8.4.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
mypy-extensions==0.4.3
    # via black
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.6.2
    # via -r requirements.in
platformdirs==2.5.2
    # via black
prompt-toolkit==3.0.29
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.12.0
    # via ipython
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
requests==2.28.0
    # via -r requirements.in
six==1.16.0
    # via asttokens
sqlparse==0.4.2
    # via django
stack-data==0.2.0
    # via ipython
tomli==2.0.1
    # via
    #   black
    #   pep517
traitlets==5.2.2.post1
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools
xmltodict==0.13.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
